Fixed: Remove obsolete XBMC HTTP notification API
