+++ src/NzbDrone.Core/Notifications/Xbmc/XbmcJsonApiProxy.cs
@@ -76,7 +76,7 @@ namespace NzbDrone.Core.Notifications.Xbmc
 
             var response = ProcessRequest(request, settings, "Player.GetActivePlayers");
 
-            return Json.Deserialize<ActivePlayersEdenResult>(response).Result;
+            return Json.Deserialize<ActivePlayersResult>(response).Result;
         }
 
         public List<KodiArtist> GetArtist(XbmcSettings settings)
