+++ src/NzbDrone.Core/Notifications/Xbmc/XbmcService.cs
@@ -1,12 +1,7 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using FluentValidation.Results;
-using Newtonsoft.Json.Linq;
 using NLog;
-using NzbDrone.Common.Cache;
-using NzbDrone.Common.Serializer;
-using NzbDrone.Core.Notifications.Xbmc.Model;
 using NzbDrone.Core.Music;
 
 namespace NzbDrone.Core.Notifications.Xbmc
@@ -22,95 +17,97 @@ namespace NzbDrone.Core.Notifications.Xbmc
     public class XbmcService : IXbmcService
     {
         private readonly IXbmcJsonApiProxy _proxy;
-        private readonly IEnumerable<IApiProvider> _apiProviders;
         private readonly Logger _logger;
 
-        private readonly ICached<XbmcVersion> _xbmcVersionCache;
-
         public XbmcService(IXbmcJsonApiProxy proxy,
-                           IEnumerable<IApiProvider> apiProviders,
-                           ICacheManager cacheManager,
                            Logger logger)
         {
             _proxy = proxy;
-            _apiProviders = apiProviders;
             _logger = logger;
-
-            _xbmcVersionCache = cacheManager.GetCache<XbmcVersion>(GetType());
         }
 
         public void Notify(XbmcSettings settings, string title, string message)
         {
-            var provider = GetApiProvider(settings);
-            provider.Notify(settings, title, message);
+            _proxy.Notify(settings, title, message);
         }
 
         public void Update(XbmcSettings settings, Artist artist)
         {
-            var provider = GetApiProvider(settings);
-            provider.Update(settings, artist);
+            if (!settings.AlwaysUpdate)
+            {
+                _logger.Debug("Determining if there are any active players on XBMC host: {0}", settings.Address);
+                var activePlayers = _proxy.GetActivePlayers(settings);
+
+                if (activePlayers.Any(a => a.Type.Equals("audio")))
+                {
+                    _logger.Debug("Audio is currently playing, skipping library update");
+                    return;
+                }
+            }
+
+            UpdateLibrary(settings, artist);
         }
 
         public void Clean(XbmcSettings settings)
         {
-            var provider = GetApiProvider(settings);
-            provider.Clean(settings);
+            _proxy.CleanLibrary(settings);
         }
 
-        private XbmcVersion GetJsonVersion(XbmcSettings settings)
+        public string GetArtistPath(XbmcSettings settings, Artist artist)
         {
-            return _xbmcVersionCache.Get(settings.Address, () =>
+            var allArtists = _proxy.GetArtist(settings);
+
+            if (!allArtists.Any())
             {
-                var response = _proxy.GetJsonVersion(settings);
-
-                _logger.Debug("Getting version from response: " + response);
-                var result = Json.Deserialize<XbmcJsonResult<JObject>>(response);
-
-                var versionObject = result.Result.Property("version");
-
-                if (versionObject.Value.Type == JTokenType.Integer)
-                {
-                    return new XbmcVersion((int)versionObject.Value);
-                }
-
-                if (versionObject.Value.Type == JTokenType.Object)
-                {
-                    return Json.Deserialize<XbmcVersion>(versionObject.Value.ToString());
-                }
-
-                throw new InvalidCastException("Unknown Version structure!: " + versionObject);
-            }, TimeSpan.FromHours(12));
-        }
-
-        private IApiProvider GetApiProvider(XbmcSettings settings)
-        {
-            var version = GetJsonVersion(settings);
-            var apiProvider = _apiProviders.SingleOrDefault(a => a.CanHandle(version));
-
-            if (apiProvider == null)
-            {
-                var message = string.Format("Invalid API Version: {0} for {1}", version, settings.Address);
-                throw new InvalidXbmcVersionException(message);
+                _logger.Debug("No Artists returned from XBMC");
+                return null;
             }
 
-            return apiProvider;
+            var matchingArtist = allArtists.FirstOrDefault(s =>
+            {
+                var musicBrainzId = s.MusicbrainzArtistId.FirstOrDefault();
+
+                return musicBrainzId == artist.Metadata.Value.ForeignArtistId || s.Label == artist.Name;
+            });
+
+            return matchingArtist?.File;
+        }
+
+        private void UpdateLibrary(XbmcSettings settings, Artist artist)
+        {
+            try
+            {
+                var artistPath = GetArtistPath(settings, artist);
+
+                if (artistPath != null)
+                {
+                    _logger.Debug("Updating artist {0} (Path: {1}) on XBMC host: {2}", artist, artistPath, settings.Address);
+                }
+
+                else
+                {
+                    _logger.Debug("Artist {0} doesn't exist on XBMC host: {1}, Updating Entire Library", artist,
+                                 settings.Address);
+                }
+
+                var response = _proxy.UpdateLibrary(settings, artistPath);
+
+                if (!response.Equals("OK", StringComparison.InvariantCultureIgnoreCase))
+                {
+                    _logger.Debug("Failed to update library for: {0}", settings.Address);
+                }
+            }
+
+            catch (Exception ex)
+            {
+                _logger.Debug(ex, ex.Message);
+            }
         }
 
         public ValidationFailure Test(XbmcSettings settings, string message)
         {
-            _xbmcVersionCache.Clear();
-
             try
             {
-                _logger.Debug("Determining version of Host: {0}", settings.Address);
-                var version = GetJsonVersion(settings);
-                _logger.Debug("Version is: {0}", version);
-
-                if (version == new XbmcVersion(0))
-                {
-                    throw new InvalidXbmcVersionException("Version received from XBMC is invalid, please correct your settings.");
-                }
-
                 Notify(settings, "Test Notification", message);
             }
             catch (Exception ex)
